Cleaned up tests a bit
Moved all tests for a given function into a single test function.
Go doesn't throw an exception out of the test entirely on failure, so why not?
It looks cleaner, but produces a more verbose error, showing all of the log messages for the test.
I'm torn on how I feel about this. I'll leave it alone for now.

I inlined all the result references because they weren't doing much.